+++ CHANGELOG.md
@@ -5,9 +5,11 @@
 ### Added
 
 - Introducing capability to set the servletUri to be used by the OAuthConfiguration
+- `Store.rawUpdateQuery(String)` allows making raw SPARQL UPDATE queries.
 
 ### Changed
 
+- TRS Client uses Lyo Store instead of using Eclipse RDF4J directly.
 - Kotlin 1.9.0 is used; `kotlin-stdlib-jdk8` dependency was replaced with `kotlin-stdlib` due to [Kotlin updates](https://kotlinlang.org/docs/whatsnew18.html#updated-jvm-compilation-target).
 - Allow application to reset the oauth token cached within the server, when it deems that it is no longer valid
 - 🧨Corrected cardinality and range of the oslc_config:acceptedBy property (from String[0..1] to Resource[0..*])
@@ -17,6 +19,7 @@
 ### Removed
 
 - 🧨 Support for JDK 11 (and all versions below 17) is removed. **JDK 17 is the new baseline for Eclipse Lyo.** The SDK and sample code has been tested using JDK 17, 20, and 21-ea.
+- TRS Client no longer depends on Eclipse RDF4J. Helper methods for RDF4J were also removed.
 
 ### Fixed